主题：STM32客户端
详细描述：
1、STM32客户端
1) 解决语音模块通过串口发消息给STM32，没现象的问题：
语音模块图形串口配置有误，没初始化GPIO；
STM32串口1中断接收函数有问题，if语句漏了花括号；
板子由外部电源供电，串口1就用不了，发现默认接到USB口了，拔掉跳线帽即可；
2) 修复蓝牙模块对接的串口2，接收中断函数接收消息有误的问题；
3) 添加3个按键，并链接到家居场景模式；
4) 重构代码编写风格。特别的是，即使if语句只有一条语句，也添加花括号。

+++ STM32客户端/智能家居/hardware/led.h
@@ -22,7 +22,7 @@ typedef enum
 
 void room_lamp_init(void);
 
-void room_lamp_adjust(uint16_t adjust_value);
+void room_lamp_adjust(int adjust_value);
 
 void room_lamp_up(void);
